requirements.txt
aiogram==3.26.0
aiosqlite==0.20.0
apscheduler==3.11.0
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.7.4
typing-extensions==4.9.0
